all Thread.sleep() removed

+++ project1/src/protocol/subprotocol/handler/RemovedHandler.java
@@ -7,10 +7,11 @@ import protocol.ChunkInfo;
 
 import java.io.File;
 import java.util.List;
+import java.util.concurrent.TimeUnit;
 
 import static protocol.subprotocol.Subprotocol.REMOVED;
 
-public class RemovedHandler extends Handler{
+public class RemovedHandler extends Handler implements Runnable {
 
     private static final int REMOVED_INBETWEEN_TIME_MS = 1000;
 
@@ -25,39 +26,44 @@ public class RemovedHandler extends Handler{
         System.out.println("protocol.subprotocol.handler.RemovedHandler.run");
         Peer.getDataContainer().setStorageCapacity(maxDiskSpace * 1000);
 
+        Peer.getExecutor().execute(this);
+
+    }
+
+    @Override
+    public void run() {
+        if (Peer.getDataContainer().getCurrStorageAmount() <= Peer.getDataContainer().getStorageCapacity())
+            return;
+
         String chunkId, fileId, pathname = Chunk.getPathname();
         File chunkFile;
         ChunkInfo chunkInfo;
 
-
         List sortedBackedUpChunks = Peer.getDataContainer().getBackedUpChunksSortedInfo();
-        for (int i = 0; i < sortedBackedUpChunks.size(); i++) {
-            if (Peer.getDataContainer().getCurrStorageAmount() <= Peer.getDataContainer().getStorageCapacity())
-                return;
 
-            chunkInfo = (ChunkInfo) sortedBackedUpChunks.get(i);
-            chunkId = Chunk.buildChunkFileId(chunkInfo.getChunkNo());
-            fileId = chunkInfo.getFileId();
-            chunkFile = new File(pathname + fileId + "/" + chunkId);
+        if(sortedBackedUpChunks.size() < 1)
+            return;
 
-            if (!chunkFile.delete())
-                continue;
+        chunkInfo = (ChunkInfo) sortedBackedUpChunks.get(0);
+        chunkId = Chunk.buildChunkFileId(chunkInfo.getChunkNo());
+        fileId = chunkInfo.getFileId();
+        chunkFile = new File(pathname + fileId + "/" + chunkId);
 
-            System.out.println(Peer.getDataContainer().getCurrStorageAmount());
-            String chunkKey = Chunk.buildChunkKey(fileId, chunkInfo.getChunkNo());
-            Peer.getDataContainer().setBackedUpChunkOnPeer(chunkKey, false);
-            byte[] message = buildMessage(REMOVED, MSG_CONFIG_REMOVED, chunkInfo.getFileId(), chunkInfo.getChunkNo(), -1, null);
-
-            try {
-                Thread.sleep(REMOVED_INBETWEEN_TIME_MS);
-            } catch (InterruptedException e) {
-                e.printStackTrace();
-            }
-
-            Peer.getControlChannel().write(message);
+        if (!chunkFile.delete()) {
+            System.out.println("Error Deleting...");
+            return;
         }
-    }
 
+        System.out.println(Peer.getDataContainer().getCurrStorageAmount());
+        String chunkKey = Chunk.buildChunkKey(fileId, chunkInfo.getChunkNo());
+        Peer.getDataContainer().setBackedUpChunkOnPeer(chunkKey, false);
+        byte[] message = buildMessage(REMOVED, MSG_CONFIG_REMOVED, chunkInfo.getFileId(), chunkInfo.getChunkNo(), -1, null);
+
+        Peer.getControlChannel().write(message);
+
+        Peer.getExecutor().schedule(this, REMOVED_INBETWEEN_TIME_MS, TimeUnit.MILLISECONDS);
+
+    }
 }
